lambda_strip_docx/pyproject.toml
[tool.poetry]
name = "lambda-strip-docx"
version = "0.1.0"
description = "AWS Lambda function for cleaning DOCX and PDF documents"
authors = ["The National Archives <digitalrecords@nationalarchives.gov.uk>"]
readme = "README.md"
packages = [{include = "lambda_strip_docx", from = "."}]

[tool.poetry.dependencies]
python = "^3.13"
lxml = "6.0.2"
rollbar = "^1.3.0"
boto3 = "^1.40.52"
pymupdf = "^1.26.5"
Pillow = "^12.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "8.4.2"
moto = {extras = ["s3"], version = "5.1.15"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev]
optional = true
